requirements.txt
absl-py==0.15.0
adversarial-robustness-toolbox==1.14.0
astor==0.8.1
astunparse==1.6.3
attrs==22.2.0
cachable==0.0.1
cached-property==1.5.2
cachetools==4.2.4
causalgraphicalmodels==0.0.4
certifi==2022.12.7
charset-normalizer==2.0.12
choix==0.3.5
clang==5.0
cycler==0.11.0
dataclasses==0.8
datalib-dev==0.0.2
decorator==4.4.2
dice-ml==0.9
eagerpy==0.30.0
flatbuffers==1.12
foolbox==3.3.3
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.18
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.19.1
grpcio==1.48.2
h5py==3.1.0
idna==3.4
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
jsonschema==3.2.0
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lime==0.2.0.1
Markdown==3.3.7
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2022.7
PyWavelets==1.1.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.15.0
smmap==5.0.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2020.9.3
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.1
typing-extensions==3.7.4.3
urllib3==1.26.13
Werkzeug==2.0.3
wrapt==1.12.1
xgboost==1.5.2
zipp==3.6.0
